+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- Change return in `Employees.all`
+- Add `Employees.find_by`
+
 ## [0.1.2]
 Fix the need to require the gem in Rails applications.